requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow==2.7
opencv-python
h5py
imgaug
IPython[all]